Moves conversion utilities to library.
Also updates some dependencies with a bugfix for `rpc.getBlockByNumber`.
