+++ package.json
@@ -14,8 +14,8 @@
 		"@types/node-fetch": "2.5.3",
 		"@zoltu/ethereum-abi-encoder": "3.2.2",
 		"@zoltu/ethereum-crypto": "2.1.1",
-		"@zoltu/ethereum-fetch-json-rpc": "10.1.0",
-		"@zoltu/ethereum-types": "8.2.2",
+		"@zoltu/ethereum-fetch-json-rpc": "10.1.1",
+		"@zoltu/ethereum-types": "8.2.3",
 		"@zoltu/solidity-typescript-generator": "3.0.2",
 		"ganache-cli": "6.7.0",
 		"jasmine": "3.5.0",
